Split object utils
