+++ src/binders/index.ts
@@ -1,5 +1,6 @@
 import { Binder, BinderRoutine } from "../interfaces/Binder";
 
+import { isNil } from "../utils/language";
 import { Collection } from "../utils/type";
 
 import $class from "./class";
@@ -28,14 +29,16 @@ export const binders: Collection<Binder | BinderRoutine> = {
   },
 
   html(el: HTMLElement, value: any): void {
-    el.innerHTML = value == null ? "" : value;
+    el.innerHTML = isNil(value) ? "" : value;
   },
 
   text(el: HTMLElement, value: any): void {
+    value = isNil(value) ? "" : value;
+
     if (el.textContent) {
-      el.textContent = value == null ? "" : value;
+      el.textContent = value;
     } else {
-      el.innerText = value == null ? "" : value;
+      el.innerText = value;
     }
   }
 };
